Update to 1.6.1

+++ net/minecraft/server/Enchantment.java
@@ -68,8 +68,8 @@ public abstract class Enchantment {
         return 0;
     }
 
-    public int a(int i, EntityLiving entityliving) {
-        return 0;
+    public float a(int i, EntityLiving entityliving) {
+        return 0.0F;
     }
 
     public boolean a(Enchantment enchantment) {
