+++ net/minecraft/server/EnchantmentThorns.java
@@ -38,7 +38,7 @@ public class EnchantmentThorns extends Enchantment {
         ItemStack itemstack = EnchantmentManager.a(Enchantment.THORNS, entityliving);
 
         if (a(i, random)) {
-            entity.damageEntity(DamageSource.a(entityliving), b(i, random));
+            entity.damageEntity(DamageSource.a(entityliving), (float) b(i, random));
             entity.makeSound("damage.thorns", 0.5F, 1.0F);
             if (itemstack != null) {
                 itemstack.damage(3, entityliving);
